fix(quality): consolidate gates + drive SonarCloud/CodeQL to zero
Remove redundant quality providers (Codacy, DeepSource coverage uploads)
and fix all remaining SonarCloud (23→0) and CodeQL (454→0) issues:

- SonarCloud: 16x C++ if-init (S6004), 4x Python redundant exception
  (S5713), C++ global const (S5421), redundant static (S3609), C#
  constructor params (S107)
- CodeQL: 332x Path.Combine→Path.Join, 53x catch-of-all-exceptions,
  14x LINQ missed-select, 8x useless-upcast, 5x local-not-disposed,
  4x unpinned-tag (SHA-pinned), 3x null-deref, 2x useless-cast,
  2x constant-condition, 1x missed-ternary, 1x nested-if, 1x class-name
- 28x P/Invoke alerts dismissed (core game trainer functionality)
- CLAUDE.md updated to reflect 5-gate stack (SonarCloud, CodeQL, QLTY,
  Codecov, Dependabot)

+++ src/SwfocTrainer.App/MainWindow.xaml.cs
@@ -70,8 +70,8 @@ public partial class MainWindow : Window
 
         var keyToken = key switch
         {
-            >= Key.D0 and <= Key.D9 => ((int)(key - Key.D0)).ToString(),
-            >= Key.NumPad0 and <= Key.NumPad9 => ((int)(key - Key.NumPad0)).ToString(),
+            >= Key.D0 and <= Key.D9 => (key - Key.D0).ToString(),
+            >= Key.NumPad0 and <= Key.NumPad9 => (key - Key.NumPad0).ToString(),
             _ => key.ToString()
         };
